### 2022-08-26 #4712 @editorialbot
[PRE REVIEW]: CLOVER: A modelling framework for sustainable community-scale energy systems
**Submitting author:** @phil-sandwell (<a href="http://orcid.org/0000-0003-1117-5095">Philip Sandwell</a>)
**Repository:** https://github.com/CLOVER-energy/CLOVER
**Branch with paper.md** (empty if default branch): publish
**Version:** v5.0.5
