### 2022-08-30 a9c71e6
Update paper.bib
Update to citation of SDG7 reference
