+++ paper.bib
@@ -2,7 +2,7 @@
   title = {{Transforming our world: the 2030 Agenda for Sustainable Development}},
   author = {{United Nations}},
   year = {2015},
-  note = {{Resolution adopted by the General Assembly on 25 September 2015, (A/RES/70/1)}}
+  note = {{Resolution adopted by the General Assembly on 25 September 2015, (A/RES/70/1). Available from \url{https://digitallibrary.un.org/record/809145}.}}
 }
 
 @software{CLOVER,
