### 2022-08-30 15d5337
Update paper.bib
Update to SDG7 reference format
