### 2022-08-30 #4712 @phil-sandwell
> @phil-sandwell : Could you please check on the reference listed under "Missing DOI" above?

This reference is for a UN resolution which doesn't have a DOI. The suggested DOI is for the correct document but as a reformatted version annex of another publication, rather than the original document.

I've reformatted the reference in this paper to cite the resolution in line with the [UN guidelines](https://ask.un.org/faq/14438), including a permalink to where it's hosted in the [UN Digital Library](https://digitallibrary.un.org/record/809145). I think this is the best solution here but please advise if you have another suggestion.
