### 2022-08-30 #4712 @phil-sandwell
> @phil-sandwell : Do you have any suggestions for reviewer? This list might be of help: https://bit.ly/joss-reviewers (please start at the bottom)

Based on their research interests (energy system modelling, optimisation, renewable energy etc.) I have three suggestions for reviewers:
- brynpickering
- fneum
- EwaGomez

All have Python as a preferred language and none of them have any active reviews at present. Please let me know if you would like any further suggestions!
