+++ src/clover/simulation/solar.py
@@ -33,7 +33,9 @@ from tqdm import tqdm
 
 from ..__utils__ import (
     BColours,
+    CleanWaterScenario,
     DesalinationScenario,
+    HotWaterScenario,
     dict_to_dataframe,
     HTFMode,
     InputFileError,
@@ -116,6 +118,7 @@ def _calculate_closed_loop_solar_thermal_output(  # pylint: disable=too-many-loc
     minigrid: Minigrid,
     num_tanks: int,
     processed_total_hw_load: Optional[pd.Series],
+    relevant_scenarios: Dict[SolarPanelType, ThermalCollectorScenario],
     resource_type: ResourceType,
     scenario: Scenario,
     solar_thermal_collectors: Union[HybridPVTPanel, SolarThermalPanel],
@@ -667,7 +670,7 @@ def _calculate_closed_loop_solar_thermal_output(  # pylint: disable=too-many-loc
 
 
 def _calculate_direct_heating_solar_thermal_output(
-    collector_system_size: int,
+    collector_system_sizes: Dict[SolarPanelType, int],
     disable_tqdm: bool,
     end_hour: int,
     irradiances: pd.Series,
@@ -675,12 +678,15 @@ def _calculate_direct_heating_solar_thermal_output(
     minigrid: Minigrid,
     num_tanks: int,
     processed_total_hw_load: Optional[pd.Series],
+    relevant_scenarios: Dict[SolarPanelType, ThermalCollectorScenario],
     resource_type: ResourceType,
-    scenario: Scenario,
-    solar_thermal_collectors: Union[HybridPVTPanel, SolarThermalPanel],
+    solar_thermal_collectors: Dict[
+        SolarPanelType, Union[HybridPVTPanel, SolarThermalPanel]
+    ],
     start_hour: int,
     temperatures: pd.Series,
     thermal_desalination_plant: Optional[ThermalDesalinationPlant],
+    thermal_scenario: Union[DesalinationScenario, HotWaterScenario],
     wind_speeds: pd.Series,
 ) -> Tuple[
     pd.DataFrame, pd.DataFrame, pd.DataFrame, pd.DataFrame, pd.DataFrame, pd.DataFrame
@@ -714,8 +720,6 @@ def _calculate_direct_heating_solar_thermal_output(
             The total hot-water load placed on the system, measured in litres per hour.
         - resource_type:
             The resource type for which the PV-T output is being determined.
-        - scenario:
-            The :class:`Scenario` being considered.
         - solar_thermal_collectors:
             The solar-thermal collector(s) to model, either a solar-thermal collector as
             a `list` of :class:`SolarThermalPanel` or :class:`HybridPVTPanel` instances.
@@ -726,6 +730,8 @@ def _calculate_direct_heating_solar_thermal_output(
             period being modelled.
         - thermal_desalination_plant:
             The thermal desalination plant being considered.
+        - thermal_scenario:
+            The :class:`DeslinationScenario` or :class:`HotWaterScenario` being considered.
         - wind_speeds:
             The :class:`pd.Series` containing wind-speed information for the time period
             being modelled.
@@ -750,8 +756,128 @@ def _calculate_direct_heating_solar_thermal_output(
 
     """
 
+    import pdb
 
-def _get_relevant_scenario(
+    pdb.set_trace()
+
+    # Determine the PV-T output if present.
+    if SolarPanelType.PV_T in solar_thermal_collectors:
+        # Determine the mass flow rate.
+        collector_mass_flow_rate: float = (
+            (
+                thermal_scenario.throughput_mass_flow_rate
+                / collector_system_sizes[SolarPanelType.PV_T]
+            )
+            if thermal_scenario.throughput_mass_flow_rate is not None
[…]
