### 2022-08-31 5ba235a
Create hot-water module
