+++ src/clover/simulation/energy_system.py
@@ -31,13 +31,12 @@ import pandas as pd  # pylint: disable=import-error
 from tqdm import tqdm
 
 from ..__utils__ import (
-    AuxiliaryHeaterType,
     BColours,
     CleanWaterMode,
-    ColdWaterSupply,
     ColumnHeader,
     dict_to_dataframe,
     DieselMode,
+    FlowRateError,
     HOURS_PER_YEAR,
     HTFMode,
     InputFileError,
@@ -57,10 +56,10 @@ from ..generation.solar import solar_degradation
 from ..load.load import compute_processed_load_profile, population_hourly
 from .__utils__ import determine_available_converters, Minigrid
 from .diesel import (
-    DieselWaterHeater,
     get_diesel_energy_and_times,
     get_diesel_fuel_usage,
 )
+from .hot_water import calculate_renewable_hw_profiles
 from .solar import calculate_solar_thermal_output
 from .storage import (
     battery_iteration_step,
@@ -316,7 +315,7 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
 
     Outputs:
         - buffer_tank_temperature:
-            The temperature of the buffer tank, measured in degrees Celcius.
+            The temperature of the buffer tank, measured in degrees Celsius.
         - buffer_tank_volume_supplied:
             The volume of buffer solution outputted by the HTF buffer tanks.
         - feedwater_sources:
@@ -324,10 +323,10 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
             system.
         - clean_water_pvt_collector_input_temperature:
             The input temperature of the HTF entering the PV-T collectors, measured in
-            degrees Celcius.
+            degrees Celsius.
         - clean_water_pvt_collector_output_temperature:
             The output temperature of HTF from the PV-T collectors, measured in degrees
-            Celcius.
+            Celsius.
         - clean_water_pvt_electric_power_per_unit:
             The electric power produced by the PV-T, in kWh, per unit of PV-T installed.
         - pvt_size:
@@ -403,9 +402,10 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
             minigrid.water_pump.throughput,
             BColours.endc,
         )
-        raise InputFileError(
-            "transmission inputs",
-            "The water pump defined is unable to meet PV-T flow requirements.",
+        raise FlowRateError(
+            "water pump",
+            f"The water pump defined, {minigrid.water_pump.name}, is unable to meet PV-T "
+            "flow requirements.",
         )
 
     elif scenario.solar_thermal and (
@@ -424,10 +424,10 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
             minigrid.water_pump.throughput,
             BColours.endc,
         )
-        raise InputFileError(
-            "transmission inputs",
-            "The water pump defined is unable to meet solar-thermal flow "
-            "requirements.",
+        raise FlowRateError(
+            "water pump",
+            f"The water pump defined, {minigrid.water_pump.name}, is unable to meet PV-T "
+            "flow requirements.",
         )
 
     logger.info("Calculating clean-water PV-T/solar-thermal performance profiles.")
@@ -652,470 +652,6 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
         required_feedwater_sources,
         thermal_desalination_electric_power_consumed,
         total_waste_produced,
-    )
-
-
-def _calculate_renewable_hw_profiles(  # pylint: disable=too-many-locals, too-many-statements
[…]
