### 2022-09-12 #4712 @fraukewiese
@brynpickering , @fneum , @EwaGomez : Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
The submission under review here is [CLOVER: A modelling framework for sustainable community-scale energy systems #4712](https://github.com/CLOVER-energy/CLOVER)
