### 2022-09-12 8fdb446
:tada: Solar-thermal direct code working
