+++ src/clover/simulation/energy_system.py
@@ -1168,9 +1168,10 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
 
     # Calculate hot-water PV-T related performance profiles.
     hot_water_pump_electric_power_consumed: pd.DataFrame  # pylint: disable=unused-variable
-    hot_water_pvt_collector_input_temperature: Optional[pd.DataFrame]
-    hot_water_pvt_collector_output_temperature: Optional[pd.DataFrame]
+    hot_water_collectors_input_temperatures: Optional[pd.DataFrame]
+    hot_water_collectors_output_temperatures: Optional[pd.DataFrame]
     hot_water_pvt_electric_power_per_unit: pd.DataFrame
+    hot_water_system_output_temperature: pd.DataFrame
     hot_water_tank_temperature: Optional[pd.DataFrame]
     hot_water_tank_volume_supplied: Optional[pd.DataFrame]
     solar_thermal_hw_fraction: Optional[pd.DataFrame]
@@ -1178,9 +1179,10 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
     (
         auxiliary_heater,  # pylint: disable=unused-variable
         hot_water_power_consumed,
-        hot_water_pvt_collector_input_temperature,
-        hot_water_pvt_collector_output_temperature,
+        hot_water_collectors_input_temperatures,
+        hot_water_collectors_output_temperatures,
         hot_water_pvt_electric_power_per_unit,
+        hot_water_system_output_temperature,
         hot_water_tank_temperature,
         hot_water_tank_volume_supplied,
         hot_water_temperature_gain,
@@ -1212,7 +1214,9 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
     )
     logger.debug(
         "Mean hot-water PV-T electric power per unit: %s",
-        np.mean(hot_water_pvt_electric_power_per_unit.values),
+        np.mean(hot_water_pvt_electric_power_per_unit.values)
+        if hot_water_pvt_electric_power_per_unit is not None
+        else "N/A",
     )
 
     # Post-process dataframes.
@@ -1249,8 +1253,10 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
         RenewableEnergySource.CLEAN_WATER_PVT: (
             clean_water_pvt_electric_power_per_unit
         ),
-        RenewableEnergySource.HOT_WATER_PVT: hot_water_pvt_electric_power_per_unit,
     }
+    if hot_water_pvt_electric_power_per_unit is not None:
+        renewables_energy_map[RenewableEnergySource.HOT_WATER_PVT] = hot_water_pvt_electric_power_per_unit
+
     renewables_energy_used_directly: pd.DataFrame
     (
         battery_storage_profile,
@@ -1686,6 +1692,44 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
         [ColumnHeader.POWER_CONSUMED_BY_ELECTRIC_DEVICES.value]
     )
 
+    # Separate out the various renewable inputs.
+    pv_energy = renewables_energy_map[RenewableEnergySource.PV]
+    clean_water_pvt_energy = renewables_energy_map[
+        RenewableEnergySource.CLEAN_WATER_PVT
+    ]
+    hot_water_pvt_energy = renewables_energy_map[RenewableEnergySource.HOT_WATER_PVT]
+    total_pvt_energy = pd.DataFrame(
+        clean_water_pvt_energy.values + hot_water_pvt_energy.values
+    )
+    total_pvt_energy.columns = pd.Index(
+        [ColumnHeader.TOTAL_PVT_ELECTRICITY_SUPPLIED.value]
+    )
+
+    # Assemble electrical outputs
[…]
