+++ src/clover/simulation/solar.py
@@ -250,6 +250,102 @@ def _volume_withdrawn_from_tank(
     return tank_supply_on, volume_supplied
 
 
+def _get_collector_output_temperatures(
+    collector_input_temperature: float,
+    irradiance: float,
+    logger: Logger,
+    pvt_collector_mass_flow_rate: Optional[float],
+    relevant_scenarios: Dict[SolarPanelType, ThermalCollectorScenario],
+    solar_thermal_collectors: List[Union[HybridPVTPanel, SolarThermalPanel]],
+    st_collector_mass_flow_rate: Optional[float],
+    temperature: float,
+    wind_speed: float,
+) -> Tuple[float, Optional[float], Optional[float], Optional[float]]:
+    """
+    Calculate the output temperatures for the solar-thermal and PV-T collectors.
+
+    Inputs:
+        - collector_input_temperature:
+            The input temperature to use for the collector system.
+        - irradiance:
+            The irradiance, in W/m^2.
+        - logger:
+            The logger to use for the run.
+        - pvt_collector_mass_flow_rate:
+            The mass flow rate through the PV-T collectors.
+        - relevant_scenarios:
+            The set relevant scenarios, keyed by panel type.
+        - solar_thermal_collectors:
+            The `list` of solar-thermal and PV-T collectors to use.
+        - st_collector_mass_flow_rate:
+            The mass flow rate through any solar-thermal collectors present.
+        - temperature:
+            The current ambient temperature.
+        - wind_speed:
+            The current wind speed.
+
+    Outputs:
+        - collector_system_output_temperature:
+            The output temperature of the collector system as a whole.
+        - fractional_electric_performance:
+            The fractional electrical performance of the PV-T collectors.
+        - st_collector_output_temperature:
+            The output temperature of the solar-thermal collectors installed, if
+            applicable.
+        - pvt_collector_output_temperature:
+            The output temperature of the PV-T collectors installed, if applicable.
+
+    """
+
+    # Determine the PV-T output if present.
+    if SolarPanelType.PV_T in solar_thermal_collectors:
+        # Calculate the output temperature map from the collector.
+        fractional_electrical_performance, pvt_output_temperature = solar_thermal_collectors[SolarPanelType.PV_T].calculate_performance(
+                temperature,
+                relevant_scenarios[SolarPanelType.PV_T].htf_heat_capacity,
+                collector_input_temperature,
+                logger,
+                pvt_collector_mass_flow_rate,
+                irradiance,
+                wind_speed,
+            )
+    else:
+        fractional_electrical_performance = None
+        pvt_output_temperature = None
+
+    # Determine the solar-thermal output if present
+    if SolarPanelType.SOLAR_THERMAL in solar_thermal_collectors:
+        # Determine the correct input temperature to use if using a series setup or a
+        # solar-thermal-only system.
+        solar_thermal_input_temperature: float = (
+            pvt_output_temperature
+            if pvt_output_temperature is not None
+            else collector_input_temperature
+        )
+
+        # Calculate the output temperature map from the collector.
+        _, solar_thermal_output_temperature: List[Tuple[Optional[float], float]] = solar_thermal_collectors[
+                SolarPanelType.SOLAR_THERMAL
+            ].calculate_performance(
+                temperature,
+                relevant_scenarios[SolarPanelType.SOLAR_THERMAL].htf_heat_capacity,
+                solar_thermal_input_temperature,
+                logger,
+                st_collector_mass_flow_rate,
+                irradiance,
[…]
