### 2022-09-14 764c029
Solar-thermal closed loop working
