+++ src/clover/simulation/solar.py
@@ -300,15 +300,18 @@ def _get_collector_output_temperatures(
     # Determine the PV-T output if present.
     if SolarPanelType.PV_T in solar_thermal_collectors:
         # Calculate the output temperature map from the collector.
-        fractional_electrical_performance, pvt_output_temperature = solar_thermal_collectors[SolarPanelType.PV_T].calculate_performance(
-                temperature,
-                relevant_scenarios[SolarPanelType.PV_T].htf_heat_capacity,
-                collector_input_temperature,
-                logger,
-                pvt_collector_mass_flow_rate,
-                irradiance,
-                wind_speed,
-            )
+        (
+            fractional_electrical_performance,
+            pvt_output_temperature,
+        ) = solar_thermal_collectors[SolarPanelType.PV_T].calculate_performance(
+            temperature,
+            relevant_scenarios[SolarPanelType.PV_T].htf_heat_capacity,
+            collector_input_temperature,
+            logger,
+            pvt_collector_mass_flow_rate,
+            irradiance,
+            wind_speed,
+        )
     else:
         fractional_electrical_performance = None
         pvt_output_temperature = None
@@ -324,22 +327,24 @@ def _get_collector_output_temperatures(
         )
 
         # Calculate the output temperature map from the collector.
-        _, solar_thermal_output_temperature: List[Tuple[Optional[float], float]] = solar_thermal_collectors[
-                SolarPanelType.SOLAR_THERMAL
-            ].calculate_performance(
-                temperature,
-                relevant_scenarios[SolarPanelType.SOLAR_THERMAL].htf_heat_capacity,
-                solar_thermal_input_temperature,
-                logger,
-                st_collector_mass_flow_rate,
-                irradiance,
-                wind_speed,
-            )
+        _, solar_thermal_output_temperature = solar_thermal_collectors[
+            SolarPanelType.SOLAR_THERMAL
+        ].calculate_performance(
+            temperature,
+            relevant_scenarios[SolarPanelType.SOLAR_THERMAL].htf_heat_capacity,
+            solar_thermal_input_temperature,
+            logger,
+            st_collector_mass_flow_rate,
+            irradiance,
+            wind_speed,
+        )
     else:
         solar_thermal_output_temperature = None
 
     return (
-        solar_thermal_output_temperature if solar_thermal_output_temperature is not None else pvt_output_temperature,
+        solar_thermal_output_temperature
+        if solar_thermal_output_temperature is not None
+        else pvt_output_temperature,
         fractional_electrical_performance,
         pvt_output_temperature,
         solar_thermal_output_temperature,
@@ -621,10 +626,6 @@ def _calculate_closed_loop_solar_thermal_output(  # pylint: disable=too-many-loc
 
     """
 
-    import pdb
-
-    pdb.set_trace()
-
     if minigrid.heat_exchanger is None:
         logger.error(
             "%sThe energy system does not contain a heat exchanger despite the %s "
@@ -710,7 +711,6 @@ def _calculate_closed_loop_solar_thermal_output(  # pylint: disable=too-many-loc
             round(pvt_collector_mass_flow_rate, 2),
         )
 
-
     else:
         raise ProgrammerJudgementFault(
             "simulation.solar::_calculate_closed_loop_solar_thermal_output",
@@ -785,12 +785,15 @@ def _calculate_closed_loop_solar_thermal_output(  # pylint: disable=too-many-loc
     collector_input_temperature_map: Dict[
         SolarPanelType, Dict[int, float]
     ] = collections.defaultdict(
-        lambda: collections.defaultdict(default_supply_temperature)
+        lambda: collections.defaultdict(lambda: default_supply_temperature)
     )
     collector_output_temperature_map: Dict[
         SolarPanelType, Dict[int, float]
[…]
