### 2022-09-14 5422105
Solar-thermal direct smooth, outputs unverified
