### 2022-09-14 d2786b7
:construction: Fractional electrical performance not saving for non times
