+++ src/clover/simulation/energy_system.py
@@ -1698,45 +1698,69 @@ def run_simulation(  # pylint: disable=too-many-locals, too-many-statements
     # Separate out the various renewable inputs.
     pv_energy = renewables_energy_map[RenewableEnergySource.PV]
 
+    # Add column headers to electric system performance outputs
+    battery_health_frame.columns = pd.Index([ColumnHeader.BATTERY_HEALTH.value])
+    blackout_times.columns = pd.Index([ColumnHeader.BLACKOUTS.value])
+    diesel_fuel_usage.columns = pd.Index([ColumnHeader.DIESEL_FUEL_USAGE.value])
+    diesel_times.columns = pd.Index([ColumnHeader.DIESEL_GENERATOR_TIMES.value])
+    energy_deficit_frame.columns = pd.Index([ColumnHeader.ELECTRICITY_DEFICIT.value])
+    energy_surplus_frame.columns = pd.Index([ColumnHeader.DUMPED_ELECTRICITY.value])
+    hourly_battery_storage_frame.columns = pd.Index(
+        [ColumnHeader.HOURLY_STORAGE_PROFILE.value]
+    )
+    households.columns = pd.Index([ColumnHeader.HOUSEHOLDS.value])
+    diesel_energy.columns = pd.Index([ColumnHeader.DIESEL_ENERGY_SUPPLIED.value])
+    kerosene_mitigation.columns = pd.Index([ColumnHeader.KEROSENE_MITIGATION.value])
+    kerosene_usage.columns = pd.Index([ColumnHeader.KEROSENE_LAMPS.value])
+    storage_power_supplied_frame.columns = pd.Index(
+        [ColumnHeader.ELECTRICITY_FROM_STORAGE.value]
+    )
+    total_energy_used.columns = pd.Index(
+        [ColumnHeader.TOTAL_ELECTRICITY_CONSUMED.value]
+    )
+    unmet_energy.columns = pd.Index([ColumnHeader.UNMET_ELECTRICITY.value])
+
     # Assemble electrical outputs
     system_performance_outputs_list = [
-        load_energy,
-        total_energy_used,
-        power_used_on_electricity,
-        unmet_energy,
+        battery_health_frame,
+        battery_storage_profile,
         blackout_times,
-        renewables_energy_used_directly,
-        storage_power_supplied_frame,
-        grid_energy,
+        diesel_fuel_usage,
         diesel_energy,
         diesel_times,
-        diesel_fuel_usage,
-        battery_storage_profile,
-        pv_energy,
-        renewables_energy,
-        hourly_battery_storage_frame,
         energy_deficit_frame,
         energy_surplus_frame,
-        battery_health_frame,
+        grid_energy,
+        hourly_battery_storage_frame,
         households,
         kerosene_usage,
         kerosene_mitigation,
+        load_energy,
+        power_used_on_electricity,
+        pv_energy,
+        renewables_energy,
+        renewables_energy_used_directly,
+        storage_power_supplied_frame,
+        total_energy_used,
+        unmet_energy,
     ]
 
     # PV-T electrical performance outputs.
     if scenario.pv_t:
-        clean_water_pvt_energy = renewables_energy_map[
+        # Determine the electricity supplied by PV-T
+        clean_water_pvt_electricity = renewables_energy_map[
             RenewableEnergySource.CLEAN_WATER_PVT
         ]
-        hot_water_pvt_energy = renewables_energy_map[
+        hot_water_pvt_electricity = renewables_energy_map[
             RenewableEnergySource.HOT_WATER_PVT
         ]
-        total_pvt_energy = pd.DataFrame(
-            clean_water_pvt_energy.values + hot_water_pvt_energy.values
+        total_pvt_electricity = pd.DataFrame(
[…]
