### 2022-10-10 6cb1fa9
Diesel generator get settings function done
