+++ src/clover/simulation/diesel.py
@@ -44,6 +44,8 @@ __all__ = (
     "DIESEL_CONSUMPTION",
     "DieselGenerator",
     "DieselWaterHeater",
+    "find_cycle_charging_threshold",
+    "get_cycle_charging_energy",
     "get_diesel_energy_and_times",
     "get_diesel_fuel_usage",
 )
@@ -79,9 +81,12 @@ class DieselGenerator:
     minimum_load: float
     name: str
 
+    capacity: Optional[float] = None
     _setting_map: Optional[Dict[int, DieselSetting]] = None
 
-    def get_setting(self, hour: int, diesel_settings: List[DieselSetting]) -> DieselSetting:
+    def get_setting(
+        self, hour: int, diesel_settings: List[DieselSetting]
+    ) -> DieselSetting:
         """
         Gets the diesel setting for the current hour and computes the mapping first time
 
@@ -109,7 +114,7 @@ class DieselGenerator:
             raise InputFileError(
                 "diesel scenario",
                 "No diesel settings were specified despite cycle charging being "
-                "requested."
+                "requested.",
             )
 
         # Compute the map
@@ -123,7 +128,10 @@ class DieselGenerator:
                         valid_settings[hour].append(diesel_setting)
                 # If the diesel settings run across the day boundary.
                 else:
-                    if hour < diesel_setting.end_hour or hour >= diesel_setting.start_hour:
+                    if (
+                        hour < diesel_setting.end_hour
+                        or hour >= diesel_setting.start_hour
+                    ):
                         valid_settings[hour].append(diesel_setting)
 
         # Raise an error if there are overlapping settings.
@@ -321,7 +329,35 @@ def _find_deficit_threshold_unmet(
     return energy_threshold
 
 
-def get_cycle_charging_energy() -> Tuple[float, float, float]:
+def find_cycle_charging_threshold(
+    previous_hour_operational: bool, setting: DieselSetting
+) -> float:
+    """
+    Calcualtes the threshold for the cycle charging for the respective hour.
+
+    Inputs:
+        - previous_hour_operational:
+            Whether the diesel generator was operating at the previous hour.
+        - setting:
+            The thresholds for the given hour, stored as a :class:`DieselSetting`
+            instance.
+
+    Outputs:
+        - The threshold value in fraction of battery charge (0.0 - 1.0).
+
+    """
+
+    if previous_hour_operational:
+        return setting.max_soc
+    return setting.min_soc
+
+
+def get_cycle_charging_energy(
+    empty_capacity: float,
+    max_diesel_energy_into_battery: float,
+    max_diesel_output: float,
+    min_diesel_energy_into_battery: float,
+) -> Tuple[float, float, float]:
     """
     Calculate cycle-charging storage parameters.
 
@@ -330,15 +366,20 @@ def get_cycle_charging_energy() -> Tuple[float, float, float]:
     maximum energy that the diesel generators can put into the batteries.
 
     Inputs:
-        - ...
+        - empty_capacity:
+            The empty capacity remaining in the batteries (kWh).
+        - max_diesel_energy_into_battery:
+            The maximum diesel energy into the battery.
+        - max_diesel_output:
+            The maximum power that the diesel generators can output.
+        - min_diesel_energy_into_battery:
+            The minimum diesel energy into the battery.
 
     Outputs:
-        - diesel_supplied:
+        - diesel_total_output:
             The total power supplied by the diesel generator.
-        - diesel_used:
-            The total power used by the diesel generator.
-        - remaining_empty_capacity:
-            The amount of remaining empty capcity within the batteries.
+        - diesel_to_battery:
+            The total power used by the diesel generator supplying the batteries.
         - surplus:
[…]
