### 2022-10-12 57258ff
:art: Black code
