### 2022-10-12 248229f
:green_heart: Tidy up some mypy errors: 80 -> 62
