+++ src/clover/simulation/energy_system.py
@@ -271,13 +271,13 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
     wind_speed_data: Optional[pd.Series],
 ) -> Tuple[
     Optional[pd.DataFrame],
+    List[Converter],
+    Dict[SolarPanelType, pd.DataFrame],
+    Dict[SolarPanelType, pd.DataFrame],
+    pd.DataFrame,
     pd.DataFrame,
     List[Converter],
-    Optional[pd.DataFrame],
-    Optional[pd.DataFrame],
     pd.DataFrame,
-    pd.DataFrame,
-    List[Converter],
     pd.DataFrame,
     Dict[WasteProduct, DefaultDict[int, float]],
 ]:
@@ -316,17 +316,15 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
     Outputs:
         - buffer_tank_temperature:
             The temperature of the buffer tank, measured in degrees Celsius.
-        - buffer_tank_volume_supplied:
-            The volume of buffer solution outputted by the HTF buffer tanks.
         - feedwater_sources:
             The :class:`Converter` instances which are a source of feedwater to the PV-T
             system.
-        - clean_water_pvt_collector_input_temperature:
-            The input temperature of the HTF entering the PV-T collectors, measured in
-            degrees Celsius.
-        - clean_water_pvt_collector_output_temperature:
-            The output temperature of HTF from the PV-T collectors, measured in degrees
-            Celsius.
+        - clean_water_collectors_input_temperatures:
+            The input temperature of the HTF entering the PV-T and solar-thermal
+            collectors, measured in degrees Celsius.
+        - clean_water_collectors_output_temperatures:
+            The output temperature of HTF from the PV-T and solar-thermal collectors,
+            measured in degrees Celsius.
         - clean_water_pvt_electric_power_per_unit:
             The electric power produced by the PV-T, in kWh, per unit of PV-T installed.
         - pvt_size:
@@ -338,6 +336,10 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
             desalination system.
         - solar_thermal_size:
             The size of the solar-thermal system installed.
+        - tank_volume_supplied:
+            The volume of buffer solution outputted by the HTF buffer tanks for
+            buffer-tank based systems, or the volume outputted by the PV-T and
+            solar-thermal collectors directly for direct-heating systems.
         - thermal_desalination_electric_power_consumed:
             The electric power consumed in operating the thermal desalination plant,
             measured in kWh.
@@ -346,6 +348,13 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
             produced by the system.
 
     """
+
+    # Raise an error if no water pump was defined.
+    if minigrid.water_pump is None:
+        raise InputFileError(
+            "energy system",
+            "No water pump was defined despite clean-water modelling being requested.",
+        )
 
     # Determine the list of available feedwater sources if relevant.
     if scenario.desalination_scenario is not None:
@@ -557,13 +566,15 @@ def _calculate_renewable_cw_profiles(  # pylint: disable=too-many-locals, too-ma
     # Compute the output of the PV-T system.
     clean_water_pvt_collector_output_temperature: Optional[pd.DataFrame]
     buffer_tank_temperature: Optional[pd.DataFrame]
+    tank_volume_supplied: pd.DataFrame
     (
-        clean_water_pvt_collector_input_temperature,
[…]
