### 2022-10-13 3d4ca7e
:green_heart: Hot-water module passes mypy
