+++ src/clover/simulation/hot_water.py
@@ -28,13 +28,16 @@ from ..__utils__ import (
     AuxiliaryHeaterType,
     BColours,
     ColdWaterSupply,
+    DesalinationScenario,
     FlowRateError,
+    HotWaterScenario,
     InputFileError,
     InternalError,
     ProgrammerJudgementFault,
     ResourceType,
     Scenario,
     SolarPanelType,
+    ThermalCollectorScenario,
     WasteProduct,
 )
 from ..conversion.conversion import Converter
@@ -46,26 +49,27 @@ __all__ = ("calculate_renewable_hw_profiles",)
 
 
 def _check_water_pump(
+    collector_scenario: Optional[ThermalCollectorScenario],
+    collector_system_size: float,
     logger: Logger,
     minigrid: Minigrid,
-    scenario: Scenario,
-    solar_thermal_size: float,
-    pvt_size: float,
 ) -> None:
     """
     Checks that the water pump is correctly sized to meet flow-rate requirements.
 
     Inputs:
+        - collector_scenario:
+            The scenario for the relevant collector being considered.
+        - collector_system_size:
+            The size of the solar-thermal or PV-T system installed, measured in
+            solar-thermal units.
         - logger:
             The logger to use for the run.
         - minigrid:
             The minigrid being simulated.
-        - scenario:
-            The current :class:`Scenario`.
-        - solar_thermal_size:
-            The size of the solar-thermal system, measured in solar-thermal units.
-        - pvt_size:
-            The size of the PV-T system, measured in PV-T units.
+
+    Outputs:
+        - Returns `None` if the water pump is sufficient to meet requirements.
 
     Raises:
         - FlowRateError:
@@ -88,66 +92,39 @@ def _check_water_pump(
             "or PV-T modelling being requested."
         )
 
-    # Determine whether the water pump is capable for supplying the PV-T or
-    # solar-thermal collectors with enough throughput.
-    if scenario.pv_t:
-        if (
-            scenario.hot_water_scenario.pvt_scenario.mass_flow_rate * pvt_size
-            > minigrid.water_pump.throughput
-        ):
-            logger.error(
-                "%sThe water pump supplied, %s, is incapable of meeting the required "
-                "PV-T flow rate of %s litres/hour. Max pump throughput: %s litres/hour."
-                "%s",
-                BColours.fail,
-                minigrid.water_pump.name,
-                scenario.hot_water_scenario.pvt_scenario.mass_flow_rate * pvt_size,
-                minigrid.water_pump.throughput,
-                BColours.endc,
-            )
-            raise FlowRateError(
-                "water pump",
-                f"The water pump defined, {minigrid.water_pump.name}, is unable to "
-                "meet PV-T flow requirements.",
-            )
+    if collector_scenario is None:
+        logger.info(
+            "No collector scenario defined so cannot check water pump flow-rate "
+            "requirements."
+        )
+        return
 
-    elif scenario.solar_thermal:
-        if (
-            scenario.hot_water_scenario.pvt_scenario.mass_flow_rate * solar_thermal_size
-            > minigrid.water_pump.throughput
-        ):
-            logger.error(
-                "%sThe water pump supplied, %s, is incapable of meeting the required "
-                "solar-thermal flow rate of %s litres/hour. Max pump throughput: %s "
-                "litres/hour.%s",
-                BColours.fail,
-                minigrid.water_pump.name,
-                scenario.hot_water_scenario.solar_thermal_scenario.mass_flow_rate
-                * pvt_size,
-                minigrid.water_pump.throughput,
-                BColours.endc,
-            )
-            raise FlowRateError(
-                "water pump",
[…]
