### 2022-10-13 9e43925
:green_heart: Mypy passes code despite direct heating.
