+++ src/clover/simulation/solar.py
@@ -256,7 +256,9 @@ def _get_collector_output_temperatures(
     logger: Logger,
     pvt_collector_mass_flow_rate: Optional[float],
     relevant_scenarios: Dict[SolarPanelType, ThermalCollectorScenario],
-    solar_thermal_collectors: List[Union[HybridPVTPanel, SolarThermalPanel]],
+    solar_thermal_collectors: Dict[
+        SolarPanelType, Union[HybridPVTPanel, SolarThermalPanel]
+    ],
     st_collector_mass_flow_rate: Optional[float],
     temperature: float,
     wind_speed: float,
@@ -299,6 +301,18 @@ def _get_collector_output_temperatures(
 
     # Determine the PV-T output if present.
     if SolarPanelType.PV_T in solar_thermal_collectors:
+        if pvt_collector_mass_flow_rate is None:
+            logger.error(
+                "%sNo PV-T mass-flow rate provided despite modelling requested.%s",
+                BColours.fail,
+                BColours.endc,
+            )
+            raise ProgrammerJudgementFault(
+                "simulation.solar::_get_collector_output_temperatures",
+                "No PV-T collector mass-flow rate when modelling PV-T collector "
+                "performance.",
+            )
+
         # Calculate the output temperature map from the collector.
         (
             fractional_electrical_performance,
@@ -326,6 +340,18 @@ def _get_collector_output_temperatures(
             else collector_input_temperature
         )
 
+        if st_collector_mass_flow_rate is None:
+            logger.error(
+                "%sNo solar-thermal mass-flow rate provided despite modelling requested.%s",
+                BColours.fail,
+                BColours.endc,
+            )
+            raise ProgrammerJudgementFault(
+                "simulation.solar::_get_collector_output_temperatures",
+                "No solar-thermal collector mass flow rate when modelling "
+                "solar-thermal collector performance.",
+            )
+
         # Calculate the output temperature map from the collector.
         _, solar_thermal_output_temperature = solar_thermal_collectors[
             SolarPanelType.SOLAR_THERMAL
@@ -341,8 +367,19 @@ def _get_collector_output_temperatures(
     else:
         solar_thermal_output_temperature = None
 
+    if solar_thermal_output_temperature is None and pvt_output_temperature is None:
+        logger.error(
+            "%sFailed to calculate either solar-thermal or PV-T output temperatures.%s",
+            BColours.fail,
+            BColours.endc,
+        )
+        raise ProgrammerJudgementFault(
+            "simulation.solar::_get_collector_output_temperatures",
+            "No output temperatures were determined as a result of the function call.",
+        )
+
     return (
-        solar_thermal_output_temperature
+        solar_thermal_output_temperature  # type: ignore [return-value]
         if solar_thermal_output_temperature is not None
         else pvt_output_temperature,
         fractional_electrical_performance,
@@ -366,12 +403,29 @@ def _get_relevant_thermal_scenario(
         - solar_thermal_collector:
             The relevant solar-thermal collector being used for the run.
 
+    Outputs:
+        The relevant thermal scenario.
+
+    Raises:
+        - ProgrammerJudgementFault:
+            Raised if the scenario could not be returned as it would have been `None`.
+
     """
 
     if resource_type == ResourceType.CLEAN_WATER:
+        if scenario.desalination_scenario is None:
+            raise ProgrammerJudgementFault(
+                "simulation.solar::_get_relevant_thermal_scenario",
[…]
