### 2022-10-13 b2e56ff
:shirt: Pylint passing bar unused and FIXME's
