+++ src/clover/simulation/hot_water.py
@@ -28,12 +28,10 @@ from ..__utils__ import (
     AuxiliaryHeaterType,
     BColours,
     ColdWaterSupply,
-    DesalinationScenario,
     FlowRateError,
     HotWaterScenario,
     InputFileError,
     InternalError,
-    ProgrammerJudgementFault,
     ResourceType,
     Scenario,
     SolarPanelType,
