+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = clover-energy
-version = 5.1.1a1
+version = 5.2.0a1
 author = Phil Sandwell, Ben Winchester and Hamish Beath
 author_email = philip.sandwell@gmail.com,benedict.winchester@gmail.com,hamishbeath@outlook.com
 description = Continuous Lifetime Optimisation of Variable Electricity Resources
