### 2022-10-13 dfc3bb6
:construction: Working on mysterious output parameters
