### 2022-10-25 CLOVER-energy/CLOVER#133 @EwaGomez
[JOSS] Paper Feedback
Your paper is easy to follow, I just have a few questions:

- There is an doi missing in your references: 10.1891/9780826190123.ap02 may be a valid DOI for the title: Transforming our world: the 2030 Agenda for Sustainable Development
- Line 137: When citing Homepages, you should also mention the date and the author. Aren't there publications about renewables.ninja?
- One author of the article did not contribute to the program. Can you specify her contribution to the paper and software?
- It is not yet clear to me who the target group of CLOVER is? Is it people living in developing countries, local companies or donors? You say "academics, practitioners and policymakers", but I don't see reasons why CLOVER is especially interesting for those groups. Maybe you could highlight that a bit more. Especially in Academia, we have access to paid tools like PVSyst or PV*Sol, which offer a larger variety of features.
- You documented CLOVER very well. I see that it is a great alternative to spreadsheet solutions, which are commonly used. Can you assure your target group has the technical knowledge to install CLOVER and use it?
- Which other libraries exist in this field and what differentiates CLOVER from them (f.e. pvlib)?
- Are you able to cite all the projects CLOVER has been used in (line 96 to 100)?

cross reference: #openjournals/joss-reviews/issues/4799
