### 2022-10-25 CLOVER-energy/CLOVER#134 @EwaGomez
[JOSS] File path in user manual
In your CLOVER user manual it is stated that you manually have to change the file path in different files. That seems very inconvenient, have you thought about automating that? In my recent PV simulation tool I used path from pathlib.
This could be an example of setting the path, where CLOVER is installed as the rootPath.

> from pathlib import Path
> rootPath = os.path.realpath(".")

It seems that the user manual is not as actual as the ReadMe, am I right? So if that might not me an issue anymore, please update the user manual.

cross-reference: #openjournals/joss-reviews/issues/4799
