### 2022-11-28 d7074b9
Update paper.bib
Amendment of reference for Renewables.ninja
