+++ paper.bib
@@ -32,6 +32,8 @@
 
 @misc{Renewables_Ninja,
   title = {{Renewables.ninja}},
+  author = {Pfenninger, Stefan and Staffell, Iain},
+  year = {2016},
   howpublished = {\url{https://www.renewables.ninja/}},
   note = {Accessed: 22 August 2022}
 }
