### 2022-11-28 8921163
Update paper.md
Update of target user groups in "Statement of need"
