+++ paper.md
@@ -45,7 +45,7 @@ CLOVER was developed to provide a robust techno-economic and climate impact anal
 
 CLOVER can provide an insight into the technical performance, costs, and greenhouse gas emissions of a system, and allow the user to evaluate many different options to decide on the best way to provide sustainable, affordable and reliable electricity to the community. Its simulation functionality offers the user an insight into the hourly, daily, seasonal and multi-year technical performance of a system in response to changing electricity demands and varying renewable resources. Its optimisation process replicates the design process common in sustainable development settings by identifying optimum system design over short-term time horizons, but allowing the system to increase in capacity over its lifetime as the demand of the community grows in response to its economic development [@Sandwell:2017a].
 
-The primary target user group of CLOVER is academic researchers who investigate electricity access in developing countries and the design of community-scale electricity systems, with examples of research outputs given in [Research and future development](#Research and future development). Secondary target groups include development practitioners and energy service providers, such as non-governmental organisations or businesses, which could use CLOVER to inform their strategies and operations. These two main target groups have collaborated to deliver academic research with a focus on the practical implementation of sustainable electricity systems in developing countries [@Beath:2021,@Few:2022,@MattheyJunod:2022].
+The primary target user group of CLOVER is academic researchers who investigate electricity access in developing countries and the design of community-scale electricity systems, with examples of research outputs given in [Research and future development](#research-and-future-development). Secondary target groups include development practitioners and energy service providers, such as non-governmental organisations or businesses, which could use CLOVER to inform their strategies and operations. These two main target groups have collaborated to deliver academic research with a focus on the practical implementation of sustainable electricity systems in developing countries [@Beath:2021;@Few:2022;@MattheyJunod:2022].
 
 # Main features
 
