### 2022-11-28 83d99c9
Update paper.md
Addition of other relevant libraries to "Statement of need"
