+++ paper.md
@@ -45,6 +45,11 @@ CLOVER was developed to provide a robust techno-economic and climate impact anal
 
 CLOVER can provide an insight into the technical performance, costs, and greenhouse gas emissions of a system, and allow the user to evaluate many different options to decide on the best way to provide sustainable, affordable and reliable electricity to the community. Its simulation functionality offers the user an insight into the hourly, daily, seasonal and multi-year technical performance of a system in response to changing electricity demands and varying renewable resources. Its optimisation process replicates the design process common in sustainable development settings by identifying optimum system design over short-term time horizons, but allowing the system to increase in capacity over its lifetime as the demand of the community grows in response to its economic development [@Sandwell:2017a].
 
+Other Python libraries exist which can be used to investigate solar and/or off-grid electricity systems, however CLOVER offers alternative or additional functionalities relevant and beneficial to the design and assess community-scale electricity systems. These alternatives include:
+* `pvlib` is an established open-source package for estimating the generation output of solar technologies [@pvlib]. It focuses on the technical performance of the PV system (modules, inverters, trackers etc.) in detail, whereas CLOVER also models and optimises other parts of the electricity system (battery storage, distribution, demand, other generation technologies) and the economic/environmental impacts.
+* `Offgridders` is an open-source Python tool for simulating and optimising minigrids [@Offgridders]. It has many of the same functions as CLOVER, however CLOVER can simulate and optimise over multiple years and resize system capacities and re-evaluate impacts as load demand grows over time, as is the aspiration in many development settings. Furthermore CLOVER can automatically generate or acquire user inputs for energy generation and demand, rather than requiring these inputs as `.csv` files.
+* `OSeMOSYS` [@OSeMOSYS] and `pypsa-earth` (developed from the established `PyPSA` framework [@PyPSA]) are both energy system models for long-term energy planning. Both have focused on applications in developing countries, similarly to CLOVER, but do so at the national or international scale for large-scale electricity networks, rather than the relatively small community-scale applications for which CLOVER is designed.
+
 The primary target user group of CLOVER is academic researchers who investigate electricity access in developing countries and the design of community-scale electricity systems, with examples of research outputs given in [Research and future development](#research-and-future-development). Secondary target groups include development practitioners and energy service providers, such as non-governmental organisations or businesses, which could use CLOVER to inform their strategies and operations. These two main target groups have collaborated to deliver academic research with a focus on the practical implementation of sustainable electricity systems in developing countries [@Beath:2021;@Few:2022;@MattheyJunod:2022].
 
 # Main features
