### 2022-11-28 2a36fdd
Update paper.bib
References for alternative packages in "Statement of need" section
