+++ paper.bib
@@ -30,6 +30,51 @@
   address = {{London, UK}}
 }
 
+@article{pvlib,
+  doi = {10.21105/joss.00884}, 
+  url = {https://doi.org/10.21105/joss.00884}, 
+  year = {2018}, 
+  publisher = {The Open Journal}, 
+  volume = {3}, 
+  number = {29}, 
+  pages = {884}, 
+  author = {William F. Holmgren and Clifford W. Hansen and Mark A. Mikofski}, 
+  title = {pvlib python: a python package for modeling solar energy systems}, 
+  journal = {Journal of Open Source Software}
+}
+
+@article{Offgridders,
+  author = {Hoffmann, M.M. and Pelz, S. and Monés-Pederzini, Ò. and Andreottola, M. and Blechinger, P.},
+  title = {{Overcoming the Bottleneck of Unreliable Grids: Increasing Reliability of Household Supply with Decentralized Backup Systems}},
+  journal = {J Sustain Res.},
+  year = {2020},
+  volume = {2},
+  issue = {1},
+  pages = {e200009},
+  doi = {10.20900/jsr20200009}
+}
+
+@software{OSeMOSYS,
+  author = {M. Howells, H. Rogner, N. Strachan, C. Heaps, H. Huntington, S. Kypreos, A. Hughes, S. Silveira, J. DeCarolis, and M. Bazillian},
+  title = {{OSeMOSYS: the open source energy modeling system: an introduction to its ethos, structure and development}},
+  journal = {Energy Policy},
+  vol = {39},
+  pages = {5850–5870},
+  year = {2011}
+}
+
+@article{PyPSA,
+   author = {T. Brown and J. H\"orsch and D. Schlachtberger},
+   title = {{PyPSA: Python for Power System Analysis}},
+   journal = {Journal of Open Research Software},
+   volume = {6},
+   issue = {1},
+   number = {4},
+   year = {2018},
+   eprint = {1707.09913},
+   doi = {10.5334/jors.188}
+}
+
 @misc{Renewables_Ninja,
   title = {{Renewables.ninja}},
   author = {Pfenninger, Stefan and Staffell, Iain},
