+++ paper.bib
@@ -54,8 +54,8 @@
   doi = {10.20900/jsr20200009}
 }
 
-@software{OSeMOSYS,
-  author = {M. Howells, H. Rogner, N. Strachan, C. Heaps, H. Huntington, S. Kypreos, A. Hughes, S. Silveira, J. DeCarolis, and M. Bazillian},
+@article{OSeMOSYS,
+  author = {Howells, M. and Rogner, H. and Strachan, N. and Heaps, C. and Huntington, H. and Kypreos, S. and Hughes, A. and Silveira, S. and DeCarolis, J. and Bazillian, M.},
   title = {{OSeMOSYS: the open source energy modeling system: an introduction to its ethos, structure and development}},
   journal = {Energy Policy},
   vol = {39},
