### 2022-11-28 a23e44d
Update paper.md
Update to "Research and future development" with previous studies that have involved CLOVER.
