+++ paper.bib
@@ -164,18 +164,18 @@
 }
 
 @article{OrtegaArriaga:2022,
-	author={Sayani, Reena and Ortega-Arriaga, Paloma and Sandwell, Philip and Babacan, Oytun and Gambhir, Ajay and Robinson, Darren and Nelson, Jenny},
-	title={{Sizing solar-based mini-grids for growing electricity demand: Insights from rural India}},
-	journal={Journal of Physics: Energy},
+  author = {Sayani, Reena and Ortega-Arriaga, Paloma and Sandwell, Philip and Babacan, Oytun and Gambhir, Ajay and Robinson, Darren and Nelson, Jenny},
+  title = {{Sizing solar-based mini-grids for growing electricity demand: Insights from rural India}},
+  journal = {Journal of Physics: Energy},
   doi = {10.1088/2515-7655/ac9dc0},
-	year={2022},
+  year={2022},
 }
 
 @misc{BarandaAlonso:2020,
   author = {Baranda Alonso, Javier and Sandwell, Philip},
   title = {{Sustainable mini-grid systems in refugee camps: A case study of Rwanda}},
   year = {2020},
-  doi = {10.25561/77296}
+  doi = {10.25561/77296},
   howpublished = {Grantham Institute, Imperial College London.}
 }
 
