+++ paper.md
@@ -76,7 +76,7 @@ CLOVER can identify the optimum energy system, defined by a selected optimisatio
 Users can specify the types of technologies available to include in the system and CLOVER uses an iterative heuristic search algorithm to identify the optimum combination and capacities to meet the needs of the community over its lifetime. A diagrammatic overview of this process is shown in @Beath:2021. 
 
 # Research and future development
-CLOVER has been used to evaluate the design and techno-economic impacts of sustainable electricity systems across a wide range of development contexts. These include the costs and greenhouse gas emissions of solar minigrids in India [@Sandwell:2017a;@Sandwell:2017b], including for healthcare applications [@Beath:2021] and as electricity demand grows over time [@OrtegaArriaga:2022]. CLOVER has been applied to community-scale electricity access in displacement settings, such as refugee camps in Rwanda [BarandaAlonso:2020;@BarandaAlonso:2021] and Djibouti [@MattheyJunod:2022], and for comparisons of the impacts of rurality and climate [@Few:2022] and existing energy infrastructure [@Sandwell:2017c] on minigrid design in different countries. To date, CLOVER has been used as part of six PhD projects and by more than 10 MSc students. 
+CLOVER has been used to evaluate the design and techno-economic impacts of sustainable electricity systems across a wide range of development contexts. These include the costs and greenhouse gas emissions of solar minigrids in India [@Sandwell:2017a;@Sandwell:2017b], including for healthcare applications [@Beath:2021] and as electricity demand grows over time [@OrtegaArriaga:2022]. CLOVER has been applied to community-scale electricity access in displacement settings, such as refugee camps in Rwanda [@BarandaAlonso:2020;@BarandaAlonso:2021] and Djibouti [@MattheyJunod:2022], and for comparisons of the impacts of rurality and climate [@Few:2022] and existing energy infrastructure [@Sandwell:2017c] on minigrid design in different countries. To date, CLOVER has been used as part of six PhD projects and by more than 10 MSc students. 
 
 CLOVER is being used as part of several ongoing research projects. These include:
 - An assessment of the cost and climate impacts of optimally-sized grid-connected solar-battery systems in India under different levels of grid availability;
