### 2022-12-04 CLOVER-energy/CLOVER#133 @phil-sandwell
Thank you @EwaGomez for the useful comments! We have addressed these in #136 with some more background below:

- This missing DOI was also mentioned in pre-review [here](https://github.com/openjournals/joss-reviews/issues/4712#issuecomment-1231770866): in summary the suggested DOI is for the correct document, but as an annex in another publication. The present citation is in line with the publishing organisation's guidelines which we think is most appropriate here.
- There are no publications introducing renewables.ninja website specifically, but we have updated the citation to one which is commonly used in the literature for this purpose adding the authors (S. Pfenninger and I. Staffell) and year (2016). The publication detailing the methods for calculating renewable generation as used in software (Pfenninger & Staffell, 2016) is included already.
- This author led and helped design the original and earlier versions of CLOVER, as well as closely supervising the development of its most recent iterations. Without her guidance the development of the software in its current state would not have been possible. She also reviewed the draft of the paper.
- We have updated the "Statement of need" section to clarify the target user groups of CLOVER. Primarily this is academic researchers who research community-scale sustainable electricity systems in developing countries (based anywhere). Secondarily this is development practitioners/NGOs/businesses working in developing countries, with examples of co-authored work between academia and practitioners given in the section.
- PVSyst and PV*Sol are useful tools but lack many of the functionalities of CLOVER, for example to generate load profiles from demand for electricity services (useful for translating fieldwork results into modelling work), add new types of PV/battery/other technologies (useful for specific academic research into novel technologies and their its applications), and to receive many more types of system data/outputs from the model (both cumulative impacts and hourly performance data over the system lifetime). This is in addition to the fact that CLOVER is open-source and fully customisable for anyone without access to paid tools.
- Members of the primary target group (academic researchers) would have the knowledge to install and use CLOVER, as evidenced by the publications from several institutions. We have also supervised 10+ MSc students from mainly (but not exclusively) engineering backgrounds who have used and/or modified CLOVER for use in their thesis projects. We are also aware of at least one minigrid company that has used CLOVER to design systems as part of their business operations. We believe that users with a relatively basic knowledge of coding (ideally in Python) can install and use CLOVER. This could come from an engineering or scientific undergraduate degree, or self-learning. More detailed applications of CLOVER or making modifications to the code would require more experience with coding but have been done by MSc students with little prior knowledge (e.g. #68).
- In the "Statement of need" section we have added brief summaries of other Python tools (`pvlib`, `Offgridders`, `OSeMOSYS`, `pypsa-earth`) and comparisons of their functionalities to that of CLOVER.
- We have reworked the "Research and future development" section to include all of the projects that CLOVER has been used on so far, as well as some ongoing projects.

We hope this addresses all of your comments, please let us know if you have any other suggestions.
