### 2022-12-04 CLOVER-energy/CLOVER#134 @phil-sandwell
Thanks very much for highlighting this. The present version of CLOVER (v5) has addressed this issue, whilst the user manual refers to the process required in v4. For new users we recommend referring to the ReadMe and Wiki as the primary sources of guidance on how to use CLOVER, rather than the old user manual which has become somewhat redundant.

To address this we have moved the old user manual to an `archive` folder (#137) which should clarify this, whilst still being available for reference if required.
