### 2022-12-04 #4799 @phil-sandwell
Thank you very much for your support with this! We've responded to the issues raised by @EwaGomez [here](https://github.com/CLOVER-energy/CLOVER/issues/133) and [here](https://github.com/CLOVER-energy/CLOVER/issues/134) and edited the submission accordingly. Please let us know if you have any more questions, comments or suggestions.
