### 2022-12-05 CLOVER-energy/CLOVER#133 @EwaGomez
Thank you @phil-sandwell - That addressed all of my issues regarding the paper and I can close this issue.
