### 2022-12-14 #4799 @EwaGomez
@fraukewiese My review is complete. The library CLOVER meets all criteria for me and I can recommend a publication in JOSS.
