### 2022-12-16 #4799 @fneum
Sorry for not getting to this earlier. I have two main comments regarding the documentation and README. Otherwise, everything looks good or has already been addressed through @EwaGomez comments.

- [ ] README: The instructions and usage examples assume that the commands are run from the root of the git repository. However, they are not applicable if you follow the `pip` route. Here, I think `python -m src.clover ....` should be replaced with `python -m clover. ....`.
- [ ] Documentation: The README is very good for getting started, but the general documentation is hidden in a PDF in the repository. A better option would be to move the documentation to a website like e.g. Readthedocs. Maybe `pandoc` can help you convert large chunks of the `.tex` to `.rst` format. Is this still up-to-date since it contains a 2020 date on the title page?
