### 2022-12-23 CLOVER-energy/CLOVER#140 @phil-sandwell
[JOSS] Paper Feedback 2
### Issue
Feedback received for the submission to JOSS. Two issues were highlighted:

1. > README: The instructions and usage examples assume that the commands are run from the root of the git repository. However, they are not applicable if you follow the `pip` route. Here, I think `python -m src.clover ....` should be replaced with `python -m clover. ....`.

2. > Documentation: The README is very good for getting started, but the general documentation is hidden in a PDF in the repository. A better option would be to move the documentation to a website like e.g. Readthedocs. Maybe `pandoc` can help you convert large chunks of the `.tex` to `.rst` format. Is this still up-to-date since it contains a 2020 date on the title page?

### Proposal
I will open a pull request to address these suggestions.
1. @BenWinchester can you look into this please?
2. I will amend the README and signpost users to the Quick start guide and Wiki as the primary documentation, and the archived PDF as a supplement.

### Note
Cross reference to the JOSS review: https://github.com/openjournals/joss-reviews/issues/4799
