+++ README.md
@@ -3,6 +3,8 @@
 CLOVER minigrid simulation and optimisation for supporting rural electrification in developing countries.
 [![DOI](https://zenodo.org/badge/476703736.svg)](https://zenodo.org/badge/latestdoi/476703736)
 
+The quick start guide below provides step-by-step introductions for downloading, setting up, and using CLOVER. For further information about CLOVER and more detailed descriptions of its functionality, please [visit the Wiki](https://github.com/CLOVER-energy/CLOVER/wiki).
+
 #### Table Of Contents
 
 [Quick start guide](#quick-start-guide)
