### 2023-01-10 CLOVER-energy/CLOVER#140 @BenWinchester
@phil-sandwell, have just got back from holiday - will take a look now. Point 1 seems like an adjustment to the :memo:README commands :smile:
