### 2023-01-24 #4799 @BenWinchester
Thank you very much for your review @fneum
An issue was raised [here](https://github.com/CLOVER-energy/CLOVER/issues/140) which was merged [here](https://github.com/CLOVER-energy/CLOVER/pull/141) with updates to address how to run the `clover-energy` package if installed via `pip` and with directions to the [wiki pages](https://github.com/CLOVER-energy/CLOVER/wiki) which recently replaced the `.pdf` userguide.
Please let me know if there's anything else that you need!
