### 2023-01-25 #4799 @fneum
Great @BenWinchester! Sorry that I didn't suspect the documentation to be placed on the Wiki Pages.

I have one mini suggestion remaining, then good to go!

In the README, it would be good to hit the keyword "documentation" somehow, in case someone is `Ctrl + F`ing for it. Either as a header or just in-text where you incorporated the reference to the wiki.
