### 2023-01-26 #4799 @fraukewiese
@BenWinchester and @phil-sandwell
Two minor points:
As raised by fneum: "In the README, it would be good to hit the keyword "documentation" somehow, in case someone is Ctrl + Fing for it. Either as a header or just in-text where you incorporated the reference to the wiki."
Please add the DOI for Howells et al 2011: https://doi.org/10.1016/j.enpol.2011.06.033
