### 2023-01-27 CLOVER-energy/CLOVER#142 @BenWinchester
Final JOSS Cleanup
### Issue
Points raised by [fraukeweise](https://github.com/fraukewiese) as part of the [submission to the journal of open-source software (JOSS)](https://github.com/openjournals/joss-reviews/issues/4799) detailed below:

> Two minor points:
> As raised by fneum: "In the README, it would be good to hit the keyword "documentation" somehow, in case someone is Ctrl + Fing for it. Either as a header or just in-text where you incorporated the reference to the wiki."
> Please add the DOI for Howells et al 2011: https://doi.org/10.1016/j.enpol.2011.06.033

### Proposal
* Add to the wiki a keyword for redirecting the documentation to the wiki pages,
* Add the missing DOI to the paper.
