### 2023-01-27 721622d
:memo: Add doi for OSeMOSYS
