+++ paper.bib
@@ -60,7 +60,8 @@
   journal = {Energy Policy},
   vol = {39},
   pages = {5850–5870},
-  year = {2011}
+  year = {2011},
+  doi = {10.1016/j.enpol.2011.06.033}
 }
 
 @article{PyPSA,
