### 2023-01-27 d8054c6
:tada: Reintroduce outbox-assembly script
