### 2023-01-27 99f2b4d
:fire: Remove JOSS paper workflow
