+++ src/clover/__main__.py
@@ -17,7 +17,7 @@ the clover module from the command-line interface.
 
 """
 
-__version__ = "5.0.5"
+__version__ = "5.0.6"
 
 import datetime
 import logging
