### 2023-01-27 #4799 @BenWinchester
@fraukewiese , the two points raised [above](https://github.com/openjournals/joss-reviews/issues/4799#issuecomment-1404784039) were addressed in [this issue](https://github.com/CLOVER-energy/CLOVER/issues/142) and then merged into our `master` branch [here](https://github.com/CLOVER-energy/CLOVER/pull/144) and published on pypi [here](https://pypi.org/project/clover-energy/5.0.6/).

- [x] _Make a tagged release of your software, and list the version tag of the archived version here._
  A tagged release, [`v5.0.6`](https://github.com/CLOVER-energy/CLOVER/releases/tag/v5.0.6), was made of the archived software,
  The tag is available [here](https://github.com/CLOVER-energy/CLOVER/releases/tag/v5.0.6);
- [x] _Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)_
  The archived software is available on Zenodo, with `v5.0.6` available at [10.5281/zenodo.7575877](https://zenodo.org/record/7575877) and the latest doi at [10.5281/zenodo.6925534](https://doi.org/10.5281/zenodo.6925534);
- [x] _Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID._
  The title of the Zenodo deposit, "CLOVER: A modelling framework for sustainable community-scale energy systems," matches that of the paper and the author list includes links to ORCID indentifiers;
- [x] _Please list the DOI of the archived version here._
  The latest DOI is **10.5281/zenodo.7575877**:
  [![DOI:10.5281/zenodo.7575877](https://zenodo.org/badge/DOI/10.5281/zenodo.7575877.svg)](https://doi.org/10.5281/zenodo.7575877)

Please let me know if there's anything else needed!
