### 2023-01-31 65883c8
Update setup.cfg
Incorrect project url in `setup.cfg`
