+++ setup.cfg
@@ -6,7 +6,7 @@ author_email = philip.sandwell@gmail.com,benedict.winchester@gmail.com,hamishbea
 description = Continuous Lifetime Optimisation of Variable Electricity Resources
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/phil-sandwell/CLOVER
+url = https://github.com/CLOVER-energy/CLOVER
 project_urls =
     Bug Tracker = https://github.com/CLOVER-energy/CLOVER/issues
 classifiers =
