### 2023-02-02 c222bba
:art: Alter `paper.md` bullet-point spacing
