+++ .github/workflows/black.yml
@@ -8,10 +8,10 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
-      - name: Set up Python 3.7.11
+      - name: Set up Python 3.7
         uses: actions/setup-python@v2
         with:
-          python-version: 3.7.11
+          python-version: 3.7.8
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
