### 2023-02-02 9961a21
:arrow_double_up: Upgrade python 3.7 :arrow_right: 3.10 in black
