### 2023-02-03 #4799 @BenWinchester
Hi @fraukewiese ,

This has now been fixed [here](https://doi.org/10.5281/zenodo.7575877) with the archive updated to match the paper.

We also spotted that bullet-point lists within the paper, which were compiling in our markdown editors, weren't being displayed properly by the action, so fixed this [here](https://github.com/CLOVER-energy/CLOVER/pull/149) . The `paper.md` is now up-to-date on both our [`publish`](https://github.com/CLOVER-energy/CLOVER/tree/publish) (stand-alone) and [`master`](https://github.com/CLOVER-energy/CLOVER/) branches. (Annoyingly, [`black`](https://github.com/psf/black/releases) and [`pylint`](https://github.com/PyCQA/pylint/releases), used for automated checking and master-branch protection, were updated in the last two days, which resulted in our automated checks failing, so there have been formatting changes to the code to pass these but no functional changes. If need be, we can trigger a new tagged release?)

Let me know if there's anything else that needs doing  :smile:
