### 2023-02-03 CLOVER-energy/CLOVER#150 @hamishbeath
Specify inverter size rather than auto sizing
### Issue
If running the model for a specific system the user may want to specify an inverter size rather than auto-sizing which is what the model currently does.

### Proposal
**Describe the solution you'd like**
An option within the model inputs that asks if you want to auto-size or not, and if you are not auto sizing, you specify an inverter size.

Leaving as is, however, for current projects there is a need to do this.
