### 2023-02-03 f83a115
:tada: Add fixed inverter size
