+++ src/clover/optimisation/__utils__.py
@@ -1028,11 +1028,7 @@ def recursive_iteration(  # pylint: disable=too-many-locals
         )
 
         # Run the simulation
-        (
-            _,
-            simulation_results,
-            system_details,
-        ) = energy_system.run_simulation(
+        (_, simulation_results, system_details,) = energy_system.run_simulation(
             int(component_sizes[RenewableEnergySource.CLEAN_WATER_PVT]),
             conventional_cw_source_profiles,
             converters,
@@ -1064,6 +1060,7 @@ def recursive_iteration(  # pylint: disable=too-many-locals
             location,
             logger,
             previous_system,
+            optimisation.scenario,
             simulation_results,
             start_year,
             system_details,
