### 2023-02-03 85c0560
:shirt: Lint ghg tests
