### 2023-02-08 #4799 @BenWinchester
Hi @kyleniemeyer , the DOI [10.5281/zenodo.7575877](doi.org/10.5281/zenodo.7575877) is correct for the reviewed version. [10.5281/zenodo.7612815](doi.org/10.5281/zenodo.7612815) is a "beta" (pre-release) version released to deal with [an issue](https://github.com/CLOVER-energy/CLOVER/issues/150) that a researcher raised with respect to their modelling, merged [here](https://github.com/CLOVER-energy/CLOVER/pull/151) and released [here](https://github.com/CLOVER-energy/CLOVER/releases/tag/v5.0.7b1) :smiley:
