### 2023-02-08 #4799 @kyleniemeyer
Congratulations @phil-sandwell & @BenWinchester et al. on your article's publication in JOSS!
